flask_app/requirements.txt
Flask==2.2.3
tensorflow==2.13.0
numpy==1.24.2
Pillow==9.5.0
Werkzeug==2.2.3
